refactor: WatchConnectiviy 통합
iOS, watchOS 의 WatchConnectiviy 를 하나로 통합하여 타겟팅했습니다.
